# Optimize revision count aggregation using ActiveRecord sum for CourseCsvBuilder#revisions_by_namespace